pyproject.toml
[project]
name = "rayna-tours-app"
version = "0.1.0"
description = "Rayna Tours ChatGPT App - MCP Server with React UI Widgets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.0",
    "cachetools>=5.5.0",
    "mcp>=1.16.0",
    "fastmcp>=2.11.3",
    "pydantic>=2.10.0",
    "starlette>=0.41.0",
    "uvicorn>=0.34.0",
    "openai>=1.30.0",
    "pinecone>=5.0.0",
    "python-dotenv>=1.0.0",
]
